plan & demo data

+++ taserver/src/main/java/com/hcit/taserver/ta/assessment/AssessmentController.java
@@ -40,7 +40,7 @@ public class AssessmentController {
   }
 
   @GetMapping("/{id}")
-  public Assessment fetch(@PathVariable Integer id) {
+  public Assessment fetch(@PathVariable Long id) {
     return assessmentService.findById(id);
   }
