+++ taserver/src/main/java/com/hcit/taserver/ta/assessment/AssessmentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AssessmentRepository extends JpaRepository<Assessment, Integer> {
+public interface AssessmentRepository extends JpaRepository<Assessment, Long> {
 
 }
